+++ Changelog.md
@@ -12,6 +12,7 @@
  - Fix running tests with test database.
  - Change data, that is seeding in production.
  - Add route to display logs.
+ - Add option to stay logged in.
 
 ## [1.1.0] - 2020-02-06
 ### Added